refactor: decompose monolithic components into testable sub-components
Split SingleSpeciesPage into BirdListTab, RetrapsTab, StatsHistoryTab,
WeightWingTab plus thin coordinator. Split StatsAccordion into
StatsAccordionItem plus shell. Split SessionHistoryCalendar into
YearGroup (owns month expansion state), MonthGroup, plus thin coordinator.

No behaviour changes. Closes #254
