+++ app/components/SessionHistoryCalendar.tsx
@@ -1,15 +1,10 @@
 'use client';
 import { type SessionWithEncountersCount } from '@/app/models/session';
-import { AccordionItem } from '@/app/components/shared/Accordion';
-import {
-	BoxyList,
-	SecondaryHeading,
-	printLocationName
-} from '@/app/components/shared/DesignSystem';
+import { printLocationName } from '@/app/components/shared/DesignSystem';
 import { useState, useEffect } from 'react';
 import { StatOutput } from './shared/StatOutput';
-import { format as formatDate } from 'date-fns';
 import { NoPrefetchLink } from './shared/NoPrefetchLink';
+import { YearGroup } from './YearGroup';
 
 function groupByDateMethod(methodName: 'getFullYear' | 'getMonth') {
 	return function (
@@ -108,83 +103,6 @@ export function SessionsByDay({
 	);
 }
 
-function SessionsOfMonth({
-	model: { viewedGroupId, monthData: month }
-}: {
-	model: { viewedGroupId: number; monthData: SessionWithEncountersCount[] };
-}) {
-	return (
-		<ol className="list-inside list-none py-3">
-			<SessionsByDay
-				sessions={month}
-				wrapperClasses="mb-2"
-				viewedGroupId={viewedGroupId}
-				dateFormat="EEEE do"
-			/>
-		</ol>
-	);
-}
-
-function MonthHeading({
-	model: { monthData: month }
-}: {
-	model: { monthData: SessionWithEncountersCount[] };
-}) {
-	return (
-		<span>
-			<span className="font-bold">
-				{formatDate(new Date(month[0].visit_date), 'MMMM')}:
-			</span>{' '}
-			{month.length} sessions,{' '}
-			{month
-				.flatMap((session) => session.encounters)
-				.reduce((acc, encounter) => acc + encounter.count, 0)}{' '}
-			birds
-		</span>
-	);
-}
-
-function YearHeading({
-	model: { yearString }
-}: {
-	model: { yearString: string };
-}) {
-	return <SecondaryHeading>{yearString}</SecondaryHeading>;
-}
-
-function MonthsOfYear({
-	model: { yearData, setExpandedMonth, expandedMonth, viewedGroupId }
-}: {
-	model: {
-		yearData: SessionWithEncountersCount[][];
-		setExpandedMonth: (month: string | false) => void;
-		expandedMonth: string | false;
-		viewedGroupId: number;
-	};
-}) {
-	return (
-		<div data-testid="months-of-year">
-			<BoxyList>
-				{yearData.map((month) => {
-					const id = formatDate(new Date(month[0].visit_date), 'yyyy-MM');
-					return (
-						<AccordionItem
-							key={id}
-							id={id}
-							HeadingComponent={MonthHeading}
-							ContentComponent={SessionsOfMonth}
-							model={{ monthData: month, viewedGroupId }}
-							onToggle={setExpandedMonth}
-							expandedId={expandedMonth}
-							icon="calendar-week"
-						/>
-					);
-				})}
-			</BoxyList>
-		</div>
-	);
-}
-
 function getYearString(year: SessionWithEncountersCount[][]) {
 	return String(new Date(year[0][0].visit_date).getFullYear());
 }
@@ -197,37 +115,27 @@ export function SessionHistoryCalendar({
 	viewedGroupId: number;
 }) {
 	const calendar = groupByYear(sessions || []).map(groupByMonth);
-	const [expandedMonth, setExpandedMonth] = useState<string | false>(false);
-	const [expandedYear, setExpandedYear] = useState(getYearString(calendar[0]));
 	const thisYearString = getYearString(calendar[0]);
+	const [expandedYear, setExpandedYear] = useState<string | false>(
+		thisYearString
+	);
+
 	useEffect(() => {
 		setExpandedYear(thisYearString);
-		setExpandedMonth(false);
 	}, [thisYearString]);
+
 	return (
 		<ol>
 			{calendar.map((year) => {
 				const yearString = getYearString(year);
 				return (
-					<AccordionItem
+					<YearGroup
 						key={yearString}
-						id={yearString}
-						testId="year-accordion-item"
-						HeadingComponent={YearHeading}
-						ContentComponent={MonthsOfYear}
-						model={{
-							yearString,
-							yearData: year,
-							setExpandedMonth,
-							expandedMonth,
-							viewedGroupId
-						}}
-						onToggle={() => {
-							setExpandedYear(yearString);
-							setExpandedMonth(false);
-						}}
-						expandedId={expandedYear}
-						icon="calendar"
+						year={year}
+						yearString={yearString}
+						viewedGroupId={viewedGroupId}
+						expandedYear={expandedYear}
+						onToggle={setExpandedYear}
 					/>
 				);
 			})}
