+++ app/components/SingleSpeciesPage.tsx
@@ -1,27 +1,20 @@
 'use client';
-import { SpeciesTable } from '@/app/components/SingleSpeciesTable';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import {
 	PageWrapper,
-	PrimaryHeading,
-	SecondaryHeading
+	PrimaryHeading
 } from '@/app/components/shared/DesignSystem';
 import { SingleSpeciesStats } from '@/app/components/SingleSpeciesStats';
 import 'chartkick/chart.js';
-import {
-	fetchPageOfBirds,
-	fetchNotableRetraps
-} from '../actions/single-species-data';
-import { useOnInView } from 'react-intersection-observer';
-import { WeightVsWingLengthChart } from '@/app/components/WeightAndWingChart';
-import { StatsHistoryChart } from '@/app/components/StatsHistoryChart';
 import type {
 	FullFatPageData,
 	PageData,
 	PageParams
 } from '@/app/(routes)/species/[speciesName]/page';
-import { NotableRetrapsTable } from './NotableRetrapsTable';
-import type { NotableRetrapsResult } from '@/app/models/db';
+import { BirdListTab } from './BirdListTab';
+import { RetrapsTab } from './RetrapsTab';
+import { StatsHistoryTab } from './StatsHistoryTab';
+import { WeightWingTab } from './WeightWingTab';
 
 function ConditionalTabPanel({
 	loadedTabs,
@@ -46,43 +39,6 @@ function ConditionalTabPanel({
 	return null;
 }
 
-function NotableRetrapsSection({
-	speciesName,
-	viewedGroupId
-}: {
-	speciesName: string;
-	viewedGroupId: number;
-}) {
-	const [notableRetraps, setNotableRetraps] = useState<NotableRetrapsResult[]>(
-		[]
-	);
-	const [isLoaded, setIsLoaded] = useState(false);
-	useEffect(() => {
-		if (notableRetraps.length > 0) return;
-		fetchNotableRetraps(speciesName, viewedGroupId).then((data) => {
-			setNotableRetraps(data);
-			setIsLoaded(true);
-		});
-	}, [speciesName, viewedGroupId, notableRetraps.length]);
-	return (
-		<>
-			<SecondaryHeading>Notable Retraps</SecondaryHeading>
-			{notableRetraps.length > 0 ? (
-				<NotableRetrapsTable data={notableRetraps} omitSpeciesName={true} />
-			) : isLoaded ? (
-				<p>No notable retraps found</p>
-			) : (
-				<div className="loading loading-spinner loading-xl"></div>
-			)}
-			{isLoaded ? null : (
-				<div className="flex items-center justify-center">
-					<div className="loading loading-spinner loading-xl"></div>
-				</div>
-			)}
-		</>
-	);
-}
-
 function SpeciesData({
 	data,
 	viewedGroupId
@@ -90,45 +46,17 @@ function SpeciesData({
 	data: FullFatPageData;
 	viewedGroupId: number;
 }) {
-	const [loadedBirds, setLoadedBirds] = useState(data.birds);
-	const [page, setPage] = useState(0);
 	const [loadedTabs, setLoadedTabs] = useState<Set<string>>(
 		new Set(['bird-list'])
 	);
-
 	const [activeTab, setActiveTab] = useState('bird-list');
+
 	function handleTabClick(event: React.MouseEvent<HTMLButtonElement>) {
 		const tab = event.currentTarget.id.replace('tabs-control-', '');
 		setLoadedTabs((prev) => new Set([...prev, tab]));
 		setActiveTab(tab);
 	}
 
-	async function loadMoreBirds() {
-		const nextPage = page + 1;
-		setPage(nextPage);
-		const newBirds = await fetchPageOfBirds(
-			data.speciesId,
-			viewedGroupId,
-			nextPage
-		);
-		setLoadedBirds([
-			...loadedBirds,
-			...newBirds.filter((bird) => !loadedIds.includes(bird.id))
-		]);
-	}
-	const isFullyLoaded =
-		loadedBirds.length >= (data.speciesStats.bird_count ?? 0);
-
-	const loadMoreRef = useOnInView(
-		(inView) => {
-			if (inView && !isFullyLoaded) {
-				loadMoreBirds();
-			}
-		},
-		{ threshold: 0 }
-	);
-
-	const loadedIds = loadedBirds.map((bird) => bird.id);
 	return (
 		<>
 			<SingleSpeciesStats {...data} viewedGroupId={viewedGroupId} />
@@ -176,23 +104,19 @@ function SpeciesData({
 				tabId="bird-list"
 				activeTabId={activeTab}
 			>
-				<SpeciesTable birds={loadedBirds} />
-				{isFullyLoaded ? null : (
-					<div
-						ref={loadMoreRef}
-						data-testid="infinite-scroll-loader"
-						className="flex items-center justify-center"
-					>
-						<div className="loading loading-spinner loading-xl"></div>
-					</div>
-				)}
+				<BirdListTab
+					speciesId={data.speciesId}
+					viewedGroupId={viewedGroupId}
+					birds={data.birds}
+					birdCount={data.speciesStats.bird_count ?? 0}
+				/>
 			</ConditionalTabPanel>
 			<ConditionalTabPanel
 				loadedTabs={loadedTabs}
 				tabId="retraps"
 				activeTabId={activeTab}
 			>
-				<NotableRetrapsSection
+				<RetrapsTab
 					speciesName={data.speciesName}
 					viewedGroupId={viewedGroupId}
 				/>
@@ -202,7 +126,7 @@ function SpeciesData({
 				tabId="stats-history"
 				activeTabId={activeTab}
 			>
-				<StatsHistoryChart
+				<StatsHistoryTab
 					speciesName={data.speciesName}
 					viewedGroupId={viewedGroupId}
 				/>
@@ -212,7 +136,7 @@ function SpeciesData({
 				tabId="size-plot"
 				activeTabId={activeTab}
 			>
-				<WeightVsWingLengthChart
+				<WeightWingTab
 					speciesId={data.speciesId}
 					viewedGroupId={viewedGroupId}
 				/>
