+++ app/components/StatsAccordion.tsx
@@ -1,14 +1,9 @@
 'use client';
 import { useState, useEffect } from 'react';
-import { AccordionItem } from './shared/Accordion';
 import { SecondaryHeading, BoxyList } from './shared/DesignSystem';
-import { StatOutput } from './shared/StatOutput';
 import type { TopPeriodsResult, TopSpeciesResult } from '@/app/models/db';
-import {
-	getTopStats,
-	type UserTopStatsArgs
-} from '@/app/actions/top-performers';
-import type { TemporalUnit } from './shared/StatOutput';
+import type { UserTopStatsArgs } from '@/app/actions/top-performers';
+import { StatsAccordionItem } from './StatsAccordionItem';
 
 export type StatConfig = {
 	id: string;
@@ -28,109 +23,6 @@ export type StatsAccordionModel = {
 	stats: AccordionItemModel[];
 };
 
-type AccordionItemModelWithGroupId = AccordionItemModel & {
-	viewedGroupId: number;
-};
-
-function hasData(data: TopPeriodsResult[] | null): data is TopPeriodsResult[] {
-	return data !== null && data.length > 0;
-}
-
-function ContentComponent({
-	model,
-	expandedId
-}: {
-	model: AccordionItemModelWithGroupId;
-	expandedId: string | false;
-}) {
-	const [data, setData] = useState<TopPeriodsResult[] | null>(model.data);
-	const [isLoading, setLoading] = useState(false);
-	const [isLoaded, setLoaded] = useState(false);
-
-	useEffect(() => {
-		if (expandedId === model.definition.id) {
-			let cancelSetLoading = false;
-			if (!isLoaded) {
-				// avoid the annoying microsecond flash of a spinner
-				setTimeout(() => {
-					if (!cancelSetLoading) {
-						setLoading(true);
-					}
-				}, 100);
-				getTopStats(Boolean(model.definition.bySpecies), {
-					...model.definition.dataArguments,
-					filters: {
-						...(model.definition.dataArguments.filters ?? {}),
-						ringing_group_filter: model.viewedGroupId
-					},
-					result_limit: 5
-				})
-					.then((data) => {
-						setData(data);
-					})
-					.catch((error) => {
-						console.error(error);
-					})
-					.finally(() => {
-						setLoaded(true);
-						cancelSetLoading = true;
-						setLoading(false);
-					});
-			}
-		}
-	}, [
-		expandedId,
-		isLoaded,
-		isLoading,
-		model.definition.id,
-		model.definition.bySpecies,
-		model.definition.dataArguments,
-		model.viewedGroupId
-	]);
-
-	return hasData(data) ? (
-		<ol className="list-inside list-none py-3">
-			{data.map((item) => (
-				<li
-					className="mb-2"
-					key={`${item.visit_date}-${(item as TopSpeciesResult).species_name} ?? ''`}
-				>
-					<StatOutput
-						value={item.metric_value}
-						speciesName={(item as TopSpeciesResult).species_name}
-						visitDate={item.visit_date}
-						showUnit={true}
-						unit={model.definition.unit}
-						temporalUnit={
-							model.definition.dataArguments.temporal_unit as TemporalUnit
-						}
-						viewedGroupId={model.viewedGroupId}
-					/>
-				</li>
-			))}
-			{isLoading && (
-				<span className="loading loading-spinner loading-xl"></span>
-			)}
-		</ol>
-	) : (
-		<span>No data available</span>
-	);
-}
-function HeadingComponent({ model }: { model: AccordionItemModelWithGroupId }) {
-	return (
-		<span>
-			<span className="font-bold">{model.definition.category}:</span>{' '}
-			{hasData(model.data) ? (
-				<span>
-					{model.data[0].metric_value} {model.definition.unit}
-				</span>
-			) : (
-				<span>No data available</span>
-			)}
-		</span>
-	);
-}
-
 export function StatsAccordion({
 	data,
 	viewedGroupId
@@ -149,14 +41,12 @@ export function StatsAccordion({
 					<SecondaryHeading>{heading}</SecondaryHeading>
 					<BoxyList>
 						{stats.map((item) => (
-							<AccordionItem
+							<StatsAccordionItem
 								key={`${viewedGroupId}-${item.definition.id}`}
-								id={item.definition.id}
-								HeadingComponent={HeadingComponent}
-								ContentComponent={ContentComponent}
-								model={{ ...item, viewedGroupId }}
+								item={item}
+								viewedGroupId={viewedGroupId}
+								expanded={expanded}
 								onToggle={setExpanded}
-								expandedId={expanded}
 							/>
 						))}
 					</BoxyList>
